refactor: reorganize section repository folders
